+++ tests/testthat/test-ModelSpecification.R
@@ -27,7 +27,7 @@ test_that("test ModelSpecification", {
     modelspec <- ModelSpecification(
       input = TunedInput(
         recipe(fo, data = df) %>%
-          step_pca(all_numeric(), -all_outcomes(), id = "pca"),
+          step_pca(all_numeric_predictors(), id = "pca"),
         grid = expand_steps(pca = list(num_comp = 6:7))
       ),
       model = TunedModel(
@@ -43,8 +43,8 @@ test_that("test ModelSpecification", {
 
     test_optim <- function(set, ...) {
       modelspec <- set(modelspec, ...)
-      mlmodel <- as.MLModel(fit(modelspec))
-      step <- mlmodel@steps[[1]]
+      model <- as.MLModel(fit(modelspec))
+      step <- model@steps[[1]]
       expect_true(step@method == modelspec@params@optim@label &&
         is(step@log$params, "tbl_df") &&
         length(step@log$params) &&